probe/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whatisup-probe"
version = "1.0.0"
description = "WhatIsUp — remote probe agent"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "WhatIsUp Contributors" }]

dependencies = [
    # HTTP checks (with redirect support and SSL)
    "httpx>=0.28.0,<0.29",

    # DNS checks
    "dnspython>=2.7.0,<3",

    # SSL certificate parsing
    "cryptography>=46.0.5,<47",

    # Scheduling
    "apscheduler>=3.11.0,<4",

    # Configuration
    "pydantic>=2.10.0,<3",
    "pydantic-settings>=2.6.0,<3",
    "python-dotenv>=1.0.1,<2",

    # Structured logging
    "structlog>=25.1.0,<26",

    # Browser scenario checks
    "playwright>=1.49.0,<2",

    # JSON Schema validation (advanced HTTP assertions)
    "jsonschema>=4.21,<5",

    # Domain WHOIS expiry checks
    "python-whois>=0.9.4,<1",

    # System health metrics (CPU, RAM, disk)
    "psutil>=6.0.0,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<10",
    "pytest-asyncio>=0.25.0,<1.4",
    "respx>=0.22.0,<0.23",
    "ruff>=0.9.0,<0.16",
    "pip-audit>=2.8.0,<3",
]

[project.scripts]
whatisup-probe = "whatisup_probe.main:main"

[tool.hatch.build.targets.wheel]
packages = ["whatisup_probe"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
